[update] fix bug in small value plotting

+++ include/SFPlot/ArcData.hpp
@@ -10,7 +10,7 @@ namespace sf
 */
 struct ArcData {
     sf::Color color; /**< color of data */
-    float value;     /**< value of data */
+    double value;     /**< value of data */
     ArcData();
 };
 };
