+++ include/SFPlot/CartesianChart.hpp
@@ -22,8 +22,8 @@ class CartesianChart : virtual public sf::Drawable, virtual public sf::Transform
 
     virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const override;
     void drawPoint(const CartesianData& elem, sf::RenderTarget& target, sf::RenderStates states,
-                   const std::pair<float, float>& x_range, const std::pair<float, float>& y_range) const;
-    void drawLine(const CartesianData& elem, sf::RenderTarget& target, sf::RenderStates states, const std::pair<float, float>& x_range, const std::pair<float, float>& y_range) const;
+                   const std::pair<double, double>& x_range, const std::pair<double, double>& y_range) const;
+    void drawLine(const CartesianData& elem, sf::RenderTarget& target, sf::RenderStates states, const std::pair<double, double>& x_range, const std::pair<double, double>& y_range) const;
 public:
     CartesianChart();
     virtual ~CartesianChart();
@@ -36,16 +36,16 @@ public:
     void setAxesColor(const sf::Color& axes_color);
     void setScaleColor(const sf::Color& axes_color);
     void autoRange();
-    void setRangeX(const float& min, const float& max);
-    void setRangeY(const float& min, const float& max);
+    void setRangeX(const double& min, const double& max);
+    void setRangeY(const double& min, const double& max);
 
     /* getter functions */
     sf::Font getFont() const;
     sf::Color getFontColor() const;
     sf::Color getAxesColor() const;
     sf::Color getScaleColor() const;
-    std::pair<float, float> getRangeX() const;
-    std::pair<float, float> getRangeY() const;
+    std::pair<double, double> getRangeX() const;
+    std::pair<double, double> getRangeY() const;
     std::shared_ptr<CartesianData> getData(const std::size_t& index);
 };
 };
