+++ include/SFPlot/FanShape.hpp
@@ -10,23 +10,23 @@ namespace sf
 */
 class FanShape : public sf::Shape
 {
-    float radius_;
-    float start_angle_;
-    float end_angle_;
+    double radius_;
+    double start_angle_;
+    double end_angle_;
 
 public:
-    FanShape(const float& radius = 50.f, const float& start_angle = 0.f, const float& end_angle = 360.f);
+    FanShape(const double& radius = 50.0, const double& start_angle = 0.0, const double& end_angle = 360.0);
     virtual ~FanShape() override;
 
     /* setter function */
-    void setRadius(const float& radius);
-    void setStartAngle(const float& start_angle);
-    void setEndAngle(const float& end_angle);
+    void setRadius(const double& radius);
+    void setStartAngle(const double& start_angle);
+    void setEndAngle(const double& end_angle);
 
     /* getter function */
-    float getRadius() const;
-    float getStartAngle() const;
-    float getEndAngle() const;
+    double getRadius() const;
+    double getStartAngle() const;
+    double getEndAngle() const;
     virtual std::size_t getPointCount() const override;
     virtual sf::Vector2f getPoint(std::size_t index) const override;
 };
