+++ include/SFPlot/RadarAxes.hpp
@@ -10,7 +10,7 @@ namespace sf
 */
 class RadarAxes : virtual public sf::Transformable, virtual public sf::Drawable
 {
-    float max_value_;
+    double max_value_;
     sf::Font font_;
     sf::Color font_color_;
     sf::Color axes_color_;
@@ -20,20 +20,20 @@ class RadarAxes : virtual public sf::Transformable, virtual public sf::Drawable
     void drawAxes(sf::RenderTarget& target, sf::RenderStates states) const;
     void drawLegend(sf::RenderTarget& target, sf::RenderStates states) const;
 public:
-    static const float RADIUS;
+    static const double RADIUS;
 
     RadarAxes();
     virtual ~RadarAxes();
 
     /* setter functions */
-    void setMaxValue(const float& max_value);
+    void setMaxValue(const double& max_value);
     void setFont(const sf::Font& font);
     void setFontColor(const sf::Color& font_color);
     void setAxesColor(const sf::Color& axes_color);
     void setScaleColor(const sf::Color& scale_color);
 
     /* getter functions */
-    float getMaxValue() const;
+    double getMaxValue() const;
     sf::Font getFont() const;
     sf::Color getFontColor() const;
     sf::Color getAxesColor() const;
