+++ src/ArcMeter.cpp
@@ -4,7 +4,7 @@
 #include "SFPlot/FanShape.hpp"
 #include "SFPlot/PlotConfig.hpp"
 
-static constexpr float ARC_MAX_ANGLE = 360.f;
+static constexpr double ARC_MAX_ANGLE = 360.0;
 namespace sf
 {
 
@@ -19,35 +19,37 @@ void ArcMeter::draw(sf::RenderTarget &target, sf::RenderStates states) const
 {
     states.transform *= getTransform();
 
-    float radius = 30.f + data_set_.size() * 20.f;
-    for(const auto& data : data_set_) {
+    double radius = 30.0 + data_set_.size() * 20.0;
+    for (const auto& data : data_set_) {
         sf::CircleShape circ(radius);
         circ.setOrigin(radius, radius);
-        circ.setPosition(50.f, 50.f);
+        circ.setPosition(50.0, 50.0);
         circ.setFillColor(sf::Color::Black);
         target.draw(circ, states);
 
         auto angle = data->value / (range_.second - range_.first) * ARC_MAX_ANGLE;
         angle = std::min(angle, ARC_MAX_ANGLE);
-        angle = std::max(angle, 0.f);
-        sf::FanShape meter(radius, 0.f, angle);
+        angle = std::max(angle, 0.0);
+        sf::FanShape meter(radius, 0.0, angle);
         meter.setOrigin(radius, radius);
-        meter.setPosition(50.f, 50.f);
+        meter.setPosition(50.0, 50.0);
         meter.setFillColor(data->color);
         target.draw(meter, states);
         radius -= 20;
     }
 
-    sf::CircleShape circ(30.f);
-    circ.setOrigin(30.f, 30.f);
-    circ.setPosition(50.f, 50.f);
+    sf::CircleShape circ(30.0);
+    circ.setOrigin(30.0, 30.0);
+    circ.setPosition(50.0, 50.0);
     circ.setFillColor(sf::Color::Black);
     target.draw(circ, states);
 
     if (!font_.getInfo().family.empty()) {
-        const float average = std::accumulate(data_set_.begin(), data_set_.end(), 0.f,
-         [](const float& a, const std::shared_ptr<ArcData>& b){return a + b->value;})
-         / data_set_.size();
+        const double average = std::accumulate(data_set_.begin(), data_set_.end(), 0.0,
+        [](const double & a, const std::shared_ptr<ArcData>& b) {
+            return a + b->value;
+        })
+        / data_set_.size();
         std::stringstream ss;
         ss.precision(2);
         ss << std::fixed;
@@ -59,7 +61,7 @@ void ArcMeter::draw(sf::RenderTarget &target, sf::RenderStates states) const
         legend.setFillColor(font_color_);
         legend.setCharacterSize(15);
         legend.setOrigin(legend.getGlobalBounds().width / 2, legend.getGlobalBounds().height / 2);
-        legend.setPosition(50.f, 50.f);
+        legend.setPosition(50.0, 50.0);
         target.draw(legend, states);
     }
 }
@@ -72,8 +74,8 @@ void ArcMeter::draw(sf::RenderTarget &target, sf::RenderStates states) const
  * @param min_range the minimal value of ArcMeter data
  * @param max_range the max value of ArcMeter data
 */
-ArcMeter::ArcMeter(const float &min_range, const float& max_range)
-    : font_(plot_config.font), font_color_(plot_config.font_color), data_set_(), range_(0.f, 100.f)
+ArcMeter::ArcMeter(const double &min_range, const double& max_range)
+    : font_(plot_config.font), font_color_(plot_config.font_color), data_set_(), range_(0.0, 100.0)
 {
 }
 
@@ -109,7 +111,7 @@ void ArcMeter::setFontColor(const sf::Color& font_color)
  * \brief set max value of plot range
  * @param max_range max value of plotting area
 */
-void ArcMeter::setMaxRange(const float &max_range)
+void ArcMeter::setMaxRange(const double &max_range)
 {
     range_.second = max_range;
     if (range_.first > range_.second)
@@ -120,7 +122,7 @@ void ArcMeter::setMaxRange(const float &max_range)
  * \brief set min value of plot range
  * @param min_range min value of plotting area
 */
-void ArcMeter::setMinRange(const float &min_range)
+void ArcMeter::setMinRange(const double &min_range)
 {
     range_.first = min_range;
     if (range_.first > range_.second)
@@ -170,7 +172,7 @@ sf::Color ArcMeter::getFontColor() const
  * \brief get plotting area value range
  * @return pair of min value and max value of plotting area
 */
-std::pair<float, float> ArcMeter::getRange() const
+std::pair<double, double> ArcMeter::getRange() const
 {
     return range_;
 }
@@ -179,7 +181,7 @@ std::pair<float, float> ArcMeter::getRange() const
  * \brief get plotting area max value
  * @return max value of plotting area
 */
-float ArcMeter::getMaxRange() const
+double ArcMeter::getMaxRange() const
 {
     return range_.second;
 }
@@ -189,7 +191,7 @@ float ArcMeter::getMaxRange() const
  * @return minimal value of plotting area
  *
 */
-float ArcMeter::getMinRange() const
+double ArcMeter::getMinRange() const
 {
     return range_.first;
 }
