+++ src/CartesianAxes.cpp
@@ -6,8 +6,8 @@
 
 namespace sf
 {
-const sf::Vector2f CartesianAxes::DIMENSION{800.f, 600.f}; /**< dimension of Cartesian chart */
-const sf::Vector2f CartesianAxes::MARGIN{40.f, 20.f}; /**< margin between plot and chart dimension */
+const sf::Vector2f CartesianAxes::DIMENSION{800.0, 600.0}; /**< dimension of Cartesian chart */
+const sf::Vector2f CartesianAxes::MARGIN{40.0, 20.0}; /**< margin between plot and chart dimension */
 
 /**
  * Drawing function derived from sf::Drawable
@@ -103,11 +103,11 @@ void CartesianAxes::drawAxes(sf::RenderTarget& target, sf::RenderStates states) 
 void CartesianAxes::drawLegend(sf::RenderTarget& target, sf::RenderStates states) const
 {
     /* Draw x legend */
-    const float x_step = (x_range_.second - x_range_.first) / 9;
-    const int x_precision = std::abs(std::min(0.f, std::log10(x_step) - 2.f));
+    const double x_step = (x_range_.second - x_range_.first) / 9;
+    const int x_precision = std::abs(std::min(0.0, std::log10(x_step) - 2.0));
 
     for (int i = 0; i < 10; i++) {
-        float val = x_range_.first + i * x_step;
+        double val = x_range_.first + i * x_step;
         std::stringstream ss;
         ss.precision(x_precision);
         ss << std::fixed;
@@ -118,16 +118,16 @@ void CartesianAxes::drawLegend(sf::RenderTarget& target, sf::RenderStates states
         legend.setFillColor(font_color_);
         legend.setCharacterSize(15);
         legend.setOrigin(legend.getGlobalBounds().width / 2.0, legend.getGlobalBounds().height / 2.0);
-        legend.setPosition(MARGIN.x + i * (DIMENSION.x - MARGIN.x) / 9, DIMENSION.y - MARGIN.y + 10.f);
+        legend.setPosition(MARGIN.x + i * (DIMENSION.x - MARGIN.x) / 9, DIMENSION.y - MARGIN.y + 10.0);
         target.draw(legend, states);
     }
 
     /* Draw y legend */
-    const float y_step = (y_range_.second - y_range_.first) / 9;
-    const int y_precision = std::abs(std::min(0.f, std::log10(y_step) - 2.f));
+    const double y_step = (y_range_.second - y_range_.first) / 9;
+    const int y_precision = std::abs(std::min(0.0, std::log10(y_step) - 2.0));
 
     for (int i = 0; i < 10; i++) {
-        float val = y_range_.first + i * y_step;
+        double val = y_range_.first + i * y_step;
         std::stringstream ss;
         ss.precision(y_precision);
         ss << std::fixed;
@@ -139,7 +139,7 @@ void CartesianAxes::drawLegend(sf::RenderTarget& target, sf::RenderStates states
         legend.setFillColor(font_color_);
         legend.setCharacterSize(15);
         legend.setOrigin(legend.getGlobalBounds().width, legend.getGlobalBounds().height);
-        legend.setPosition(MARGIN.x - 10.f, DIMENSION.y - MARGIN.y - i * (DIMENSION.y - MARGIN.y ) / 9);
+        legend.setPosition(MARGIN.x - 10.0, DIMENSION.y - MARGIN.y - i * (DIMENSION.y - MARGIN.y ) / 9.0);
         target.draw(legend, states);
     }
 }
@@ -170,7 +170,7 @@ CartesianAxes::~CartesianAxes()
  * \brief set X value of plot range
  * @param x_range X value range of plotting area
 */
-void CartesianAxes::setRangeX(const std::pair<float, float>& x_range)
+void CartesianAxes::setRangeX(const std::pair<double, double>& x_range)
 {
     x_range_ = x_range;
 }
@@ -179,7 +179,7 @@ void CartesianAxes::setRangeX(const std::pair<float, float>& x_range)
  * \brief set Y value of plot range
  * @param y_range Y value range of plotting area
 */
-void CartesianAxes::setRangeY(const std::pair<float, float>& y_range)
+void CartesianAxes::setRangeY(const std::pair<double, double>& y_range)
 {
     y_range_ = y_range;
 }
@@ -230,7 +230,7 @@ void CartesianAxes::setFontColor(const sf::Color& font_color)
  * \brief get X value of plot range
  * @return X value range of plotting area
 */
-std::pair<float, float> CartesianAxes::getRangeX() const
+std::pair<double, double> CartesianAxes::getRangeX() const
 {
     return x_range_;
 }
@@ -239,7 +239,7 @@ std::pair<float, float> CartesianAxes::getRangeX() const
  * \brief get Y value of plot range
  * @return Y value range of plotting area
 */
-std::pair<float, float> CartesianAxes::getRangeY() const
+std::pair<double, double> CartesianAxes::getRangeY() const
 {
     return y_range_;
 }
