+++ src/CartesianData.cpp
@@ -29,18 +29,18 @@ CartesianData::~CartesianData()
  * @return min, max pair of X value range
  *
 */
-std::pair<float, float> CartesianData::getRangeX() const
+std::pair<double, double> CartesianData::getRangeX() const
 {
     /* Extract Data*/
-    auto min_x = std::numeric_limits<float>::infinity();
-    auto max_x = std::numeric_limits<float>::lowest();
+    auto min_x = std::numeric_limits<double>::infinity();
+    auto max_x = std::numeric_limits<double>::lowest();
 
     for (const auto& elem : data) {
         min_x = std::min(min_x, elem.x);
         max_x = std::max(max_x, elem.x);
     }
 
-    return std::pair<float, float>(min_x, max_x);
+    return std::pair<double, double>(min_x, max_x);
 }
 
 /**
@@ -48,17 +48,17 @@ std::pair<float, float> CartesianData::getRangeX() const
  * @return min, max pair of Y value range
  *
 */
-std::pair<float, float> CartesianData::getRangeY() const
+std::pair<double, double> CartesianData::getRangeY() const
 {
     /* Extract Data*/
-    auto min_y = std::numeric_limits<float>::infinity();
-    auto max_y = std::numeric_limits<float>::lowest();
+    auto min_y = std::numeric_limits<double>::infinity();
+    auto max_y = std::numeric_limits<double>::lowest();
 
     for (const auto& elem : data) {
         min_y = std::min(min_y, elem.y);
         max_y = std::max(max_y, elem.y);
     }
 
-    return std::pair<float, float>(min_y, max_y);
+    return std::pair<double, double>(min_y, max_y);
 }
 };
